+++ firmware/native/micropython.cmake
@@ -3,6 +3,7 @@ add_library(usermod_iotmd_crypto INTERFACE)
 target_sources(usermod_iotmd_crypto INTERFACE
     ${CMAKE_CURRENT_LIST_DIR}/iotmd_crypto.c
     ${CMAKE_CURRENT_LIST_DIR}/iotmd_platform.c
+    ${CMAKE_CURRENT_LIST_DIR}/iotmd_platform_v3.c
 )
 
 target_include_directories(usermod_iotmd_crypto INTERFACE