+++ v3/__init__.py
@@ -1,0 +1,1 @@
+"""Isolated IoT-MD v3 alpha namespace."""